+++ src/main/java/ru/practicum/shareit/user/service/UserService.java
@@ -1,7 +1,6 @@
 package ru.practicum.shareit.user.service;
 
 import ru.practicum.shareit.user.dto.UserDto;
-
 import java.util.List;
 
 public interface UserService {
@@ -9,11 +8,11 @@ public interface UserService {
 
     List<UserDto> getAllUsers();
 
-    UserDto updateUser(UserDto userDto, int userId);
+    UserDto updateUser(UserDto userDto, long userId);
 
-    UserDto getUserById(int userId);
+    UserDto getUserById(long userId);
 
-    UserDto deleteUser(int userId);
+    UserDto deleteUser(long userId);
 
-    void checkOwner(int userId);
+    void checkOwner(long userId);
 }
